refactor: Folder structure and route navigation
