+++ react-auth/src/components/Head.tsx
