+++ react-auth/src/pages/Delete.tsx
