+++ react-auth/src/pages/Register.tsx
@@ -2,8 +2,8 @@ import React, { FC, useState } from "react";
 import { Form, Button, Container, Spinner } from "react-bootstrap";
 import axios from "axios";
 import Cookies from "universal-cookie";
-import { AlertCard } from "./AlertCard";
-import { Head } from "./Head";
+import { AlertCard } from "../components/AlertCard";
+import { Head } from "../components/Head";
 
 const cookies = new Cookies();
